+++ src/main.ts
@@ -2,8 +2,8 @@ import { createApp } from 'vue';
 
 import App from './App.vue';
 import { publicEnvironment } from './app/configuration/environment';
+import './design-system/theme/index.css';
 import { router } from './router';
-import './assets/base.css';
 
 void publicEnvironment;